Refactored, move save handlers to file save, renamed instances to save + added fixes to route params.
